+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Weights & Biases `TorchTrainCallback` now updates the best validation metric score as a summary statistic.
+- Added `StopEarlyCallback`, a `TorchTrainCallback` for early stopping.
 - Added support for PEP 604 in `FromParams`, i.e. writing union types as "X | Y" instead of "Union[X, Y]".
 - [internals] Added a spot for miscellaneous end-to-end integration tests (not to be confused with "tests of integrations") in `tests/end_to_end/`.
 - [internals] Core tests now run on all officially supported Python versions.
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Minor changes to `TorchTrainCallback` interface.
+- Weights & Biases `TorchTrainCallback` now updates the best validation metric score as a summary statistic.
 
 ### Fixed